Split definitions into syscore and common datasets
Also keeping docs for now.
