+++ datasets.py
@@ -1,189 +1,26 @@
 import os
-import re
-from rdflib import Graph, ConjunctiveGraph, RDF, Namespace
-from lxltools.datacompiler import Compiler, last_modified_ms
-from lxltools.timeutil import w3c_dtz_to_ms
-from lxltools.contextmaker import make_context, add_overlay
+from rdflib import Graph
+from lxltools.datacompiler import Compiler
 from urllib.parse import urljoin
 
-
-NS_PREF_ORDER = (
-        'bf2 bflc bf madsrdf skos dc dctype prov sdo bibo foaf void ldp '
-        'owl rdfs rdf xsd edtf').split()
-
-SCRA = Namespace("http://purl.org/net/schemarama#")
 
 BASE = "https://id.kb.se/"
 
 SCRIPT_DIR = os.path.dirname(__file__) or '.'
 
 
-def decorate(items, template):
-    def decorator(item):
-        for k, tplt in template.items():
-            item[k] = tplt.format(**item)
-        return item
-    return list(map(decorator, items))
-
-
-def to_camel_case(label):
-    return "".join((s[0].upper() if i else s[0].lower()) + s[1:]
-            for (i, s) in enumerate(re.split(r'[\s,.-]', label)) if s)
-
-
-def _get_repo_version():
-    try:
-        with os.popen(
-                '(cd {} && git describe --tags)'.format(SCRIPT_DIR)
-                ) as pipe:
-            return pipe.read().rstrip()
-    except:
-        return None
+def aslist(o):
+    return o if isinstance(o, list) else [] if o is None else [o]
 
 
 compiler = Compiler(base_dir=SCRIPT_DIR,
-                    dataset_id=BASE + 'dataset/definitions',
+                    dataset_id=BASE + 'dataset/common',
                     created='2013-10-17T14:07:48.000Z',
-                    tool_id=BASE + 'generator/definitions',
+                    tool_id=BASE + 'generator/datasetcompiler',
                     context='sys/context/base.jsonld',
                     record_thing_link='mainEntity',
                     system_base_iri='',
-                    union='definitions.jsonld.lines')
-
-
-def _insert_record(graph, created_ms, dataset_id):
-    entity = graph[0]
-    record = {'@type': 'Record'}
-    record[compiler.record_thing_link] = {'@id': entity['@id']}
-    graph.insert(0, record)
-    record['@id'] = compiler.generate_record_id(created_ms, entity['@id'])
-    record['inDataset'] = [{'@id': compiler.dataset_id}, {'@id': dataset_id}]
-
-
-#@compiler.dataset
-#def datasets():
-#    graph = Graph().parse(compiler.path('source/index.ttl'), format='turtle')
-#    return to_jsonld(graph, (None, None), { "@language": "sv"})
-
-
-@compiler.handler
-def vocab():
-    vocab_base = BASE + 'vocab/'
-
-    graph = compiler.construct(sources=[
-            {
-                'source': Graph().parse(str(compiler.path('source/vocab/bf-map.ttl')), format='turtle'),
-                'dataset': '?'
-            },
-            {"source": "http://id.loc.gov/ontologies/bibframe/"}
-        ],
-        query="source/vocab/bf-to-kbv-base.rq")
-
-    for part in compiler.path('source/vocab').glob('**/*.ttl'):
-        try:
-            graph.parse(str(part), format='turtle')
-        except:
-            print(f"Error in file: {part}")
-            raise
-
-    #cg = ConjunctiveGraph()
-    #cg.parse(str(compiler.path('source/vocab/display.jsonld')), format='json-ld')
-    #graph += cg
-
-    graph.update(compiler.path('source/vocab/update.rq').read_text('utf-8'))
-    graph.update(compiler.path('source/vocab/updateResource.rq').read_text('utf-8'))
-
-    rq = compiler.path('source/vocab/construct-enum-restrictions.rq').read_text('utf-8')
-    graph += Graph().query(rq).graph
-
-    rq = compiler.path('source/vocab/check-bases.rq').read_text('utf-8')
-    checks = graph.query(rq).graph
-    for check, msg in checks.subject_objects(SCRA.message):
-        print("{}: {}".format(
-            graph.qname(checks.value(check, RDF.type)).split(':')[-1],
-            msg.format(*[imp.n3() for imp in
-                checks.collection(checks.value(check, SCRA.implicated))])
-            ))
-
-    for part in compiler.path('source/marc').glob('**/*.ttl'):
-        graph.parse(str(part), format='turtle')
-
-    graph.parse(str(compiler.path('source/swepub/vocab.ttl')), format='turtle')
-
-    # Clean up generated prefixes
-    preferred = {}
-    defaulted = {}
-    for pfx, uri in graph.store.namespaces():
-        if pfx.startswith('default'):
-            defaulted[uri] = pfx
-        else:
-            preferred[uri] = pfx
-    for default_pfx, uri in defaulted.items():
-        if uri in preferred:
-            graph.namespace_manager.bind(preferred[uri], uri, override=True)
-
-    data = compiler.to_jsonld(graph)
-    del data['@context']
-
-    # Put /vocab/* first (ensures that /marc/* comes after)
-    data['@graph'] = sorted(data['@graph'], key=lambda node:
-            (not node.get('@id', '').startswith(vocab_base), node.get('@id')))
-
-    vocab_created_ms = w3c_dtz_to_ms("2013-12-31T23:00:00.000Z")
-
-    vocab_ds_url = urljoin(compiler.dataset_id, 'vocab')
-    vocab_modified_ms = last_modified_ms(compiler.current_ds_resources)
-    compiler._create_dataset_description(vocab_ds_url, vocab_created_ms, vocab_modified_ms)
-
-    _insert_record(data['@graph'], vocab_created_ms, vocab_ds_url)
-    vocab_node = data['@graph'][1]
-    version = _get_repo_version()
-    if version:
-        vocab_node['version'] = version
-
-    gen_context_path = 'sys/context/auto'
-    lib_context = make_context(graph, vocab_base, NS_PREF_ORDER)
-    add_overlay(lib_context, compiler.load_json('sys/context/kbv.jsonld'))
-    lib_context['@graph'] = [
-        {'@id': BASE + gen_context_path, '@type': 'jsonld:Context'}
-    ]
-    _insert_record(lib_context['@graph'], vocab_created_ms, vocab_ds_url)
-
-    display = compiler.load_json('source/vocab/display.jsonld')
-    _insert_record(display['@graph'], vocab_created_ms, vocab_ds_url)
-
-    compiler.write(data, "vocab")
-    compiler.write(lib_context, gen_context_path)
-    compiler.write(display, 'vocab/display')
-
-
-@compiler.handler
-def contexts():
-    contexts_ds_url = urljoin(compiler.dataset_id, 'sys/context')
-
-    context_alias = BASE + 'vocab/context'
-
-    docpath = compiler.path('sys/context/kbv.jsonld')
-    uripath = BASE + 'sys/context/kbv'
-    _write_context_record(compiler, docpath, uripath, contexts_ds_url, context_alias)
-
-    root = compiler.path('')
-    for docpath in compiler.path('sys/context').glob('target/*.jsonld'):
-        uripath = BASE + str(docpath.relative_to(root).with_suffix(''))
-        _write_context_record(compiler, docpath, uripath, contexts_ds_url)
-
-
-def _write_context_record(compiler, filepath, uripath, ds_url, alias=None):
-    ctx_data = compiler.load_json(filepath)
-    ctx_created_ms = w3c_dtz_to_ms(ctx_data.pop('created'))
-
-    ctx_data['@graph'] = [{"@id": uripath, "@type": "jsonld:Context"}]
-    _insert_record(ctx_data['@graph'], ctx_created_ms, ds_url)
-    if alias:
-        ctx_data['@graph'][0]['sameAs'] = [{'@id': alias}]
-
-    assert uripath.startswith(BASE)
-    compiler.write(ctx_data, uripath.replace(BASE, ''))
+                    union='datasets.jsonld.lines')
 
 
 @compiler.dataset
@@ -222,12 +59,6 @@ def rdaterms():
 
     return "/term/rda/", "2018-05-16T06:18:01.337Z", graph
 
-
-@compiler.dataset
-def enumterms():
-    graph = Graph().parse(str(compiler.path('source/kbv-enums.ttl')), format='turtle')
-
-    return "/term/enum/", "2018-05-29T12:36:01.337Z", graph
 
 @compiler.dataset
 def materials():
@@ -279,27 +110,6 @@ def tacnotationterms():
 
     return "/term/rda/tacnotation/", "2021-05-21T23:59:10.456Z", graph
 
-#NOTE: More suitable name might be needed if usage is broader than digital representations
-@compiler.dataset
-def reprterms():
-    graph = Graph().parse(str(compiler.path('source/repr-terms.ttl')), format='turtle')
-
-    return "/term/repr/", "2021-02-22T10:32:01.337Z", graph
-
-@compiler.dataset
-def encodingFormatterms():
-    graph = Graph().parse(str(compiler.path('source/encodingFormat-terms.ttl')), format='turtle')
-
-    return "/encodingFormat/", "2021-03-04T10:12:09.921Z", graph
-
-@compiler.dataset
-def bibdbterms():
-    graph = ConjunctiveGraph()
-    # graph = Graph()
-    for part in compiler.path('source/bibdb').glob('**/*.ttl'):
-        graph.parse(str(part), format='turtle')
-
-    return "/term/bibdb/", "2021-09-20T08:13:50.570Z", graph
 
 @compiler.dataset
 def swepubterms():
@@ -312,34 +122,6 @@ def swepubterms():
     graph.update(compiler.path('source/swepub/update.rq').read_text('utf-8'))
 
     return "/term/swepub/", "2018-05-29T12:36:01.337Z", graph
-
-
-@compiler.dataset
-def policies():
-    graph = Graph().parse(str(compiler.path('source/policies.ttl')), format='turtle')
-
-    return "/policy/", "2021-11-18T11:48:51Z", graph
-
-
-@compiler.dataset
-def containers():
-    graph = Graph().parse(str(compiler.path('source/containers.ttl')), format='turtle')
-
-    return "/term/", "2019-07-11T13:04:17.964Z", graph
-
-
-@compiler.dataset
-def generators():
-    graph = Graph().parse(str(compiler.path('source/generators.ttl')), format='turtle')
-
-    return "/generator/", "2018-04-25T18:55:14.723Z", graph
-
-
-@compiler.dataset
-def schemes():
-    graph = Graph().parse(str(compiler.path('source/schemes.ttl')), format='turtle')
-
-    return "/", "2014-02-01T20:00:01.766Z", graph
 
 
 @compiler.dataset
@@ -412,43 +194,6 @@ def countries():
         query="source/construct-countries.rq")
 
     return "/country/", "2014-02-01T12:21:14.008Z", graph
-
-
-@compiler.dataset
-def nationalities():
-    graph = Graph().parse(str(compiler.path('source/nationalities.ttl')), format='turtle')
-
-    return "/nationality/", "2014-02-01T13:08:56.596Z", graph
-
-
-@compiler.dataset
-def docs():
-    import markdown
-    docs = []
-    sourcepath = compiler.path('source')
-    for fpath in (sourcepath / 'doc').glob('**/*.mkd'):
-        text = fpath.read_text('utf-8')
-        html = markdown.markdown(text)
-        doc_id = (str(fpath.relative_to(sourcepath))
-                  .replace(os.sep, '/')
-                  .replace('.mkd', ''))
-        doc_id, dot, lang = doc_id.partition('.')
-        doc = {
-            "@type": "Article",
-            "@id": BASE + doc_id,
-            "articleBody": html
-        }
-        h1end = html.find('</h1>')
-        if h1end > -1:
-            doc['title'] = html[len('<h1>'):h1end]
-        if lang:
-            doc['language'] = {"langTag": lang},
-        docs.append(doc)
-
-    return "/doc", "2016-04-15T14:43:38.072Z", {
-        "@context": "../sys/context/base.jsonld",
-        "@graph": docs
-    }
 
 
 if __name__ == '__main__':
